Z160 has problems with surface heights less than 32 pixels so don't go above that But as long as the stride is padded to 32 bytes, it will happily render any width we like. So reduce it to 8, so we keep any pixmaps 8x32 or higher in graphics memory instead of palming them off to system memory.
We may be doing this all wrong and causing some serious memory fragmentation in
GSL though, so a much bigger width and height could be better. This gives us right
now the most stable graphics and at least the most number of accelerated pixmaps.

Better solution: only allocate pixmaps of a certain size in video memory, but also
even if in video memory, only accelerate blits of a certain size or above even if
the pixmap is rather large (after all, a 15x15 blit inside a 1280x720 screen is
not worth setting up the GPU to perform, if it is at all possible to realize that
this is the intent of a prepared composite or copy operation. Regardless, getting
the GPU to commit to an operation by finalizing registers for src,dst,alpha and
son on may still take longer including waiting for a pipeline flush than just
doing it in the CPU anyway and discarding any GPU setup we already performed).